requirements.txt
cairocffi==1.2.0
CairoSVG==2.5.2
cffi==1.14.5
cssselect2==0.4.1
defusedxml==0.7.1
lxml==4.6.3
Pillow==8.2.0
pycparser==2.20
PyQRCode==1.2.1
python-docx==0.8.11
PyYAML==5.4.1
svgwrite==1.4.1
tinycss2==1.1.0
webencodings==0.5.1
